# Tackle performance and accuracy regression of sentence tokenizer since NLTK 3.6.6